requirements.txt
# Core Dependencies
rich>=10.12.0
colorama>=0.4.4
questionary>=1.10.0
PyYAML>=6.0.0
blessed>=1.19.0
prompt-toolkit>=3.0.20

# System Monitoring
psutil>=5.8.0
py-cpuinfo>=8.0.0
netifaces>=0.11.0
distro>=1.7.0

# Security and Sandboxing
cryptography>=41.0.0
pyOpenSSL>=23.2.0
certifi>=2023.5.7
python-daemon>=3.0.1
pyinotify>=0.9.6
pyseccomp>=0.1.2
python-prctl>=1.8.1

# Performance and Hardware
python-xlib>=0.33
screeninfo>=0.8
wcwidth>=0.2.6
numpy>=1.21.0

# Tool Integration
gitpython>=3.1.30
python-crontab>=2.7.1
systemd-python>=235
libtmux>=0.18.0

# Accessibility
pyttsx3>=2.90
pyaccessibility>=0.1.0

# Async Support
asyncio>=3.4.3
aiofiles>=0.8.0
aioconsole>=0.3.3
aiohttp>=3.8.1

# Terminal Features
pyte>=0.8.1
termcolor>=2.1.0
terminaltables>=3.1.10
termios-mux>=0.1.0
blessings>=1.7
asciimatics>=1.13.0

# UI and Animation
tqdm>=4.65.0
alive-progress>=3.1.0
spinners>=0.0.24
termgraph>=0.5.3
colorthief>=0.2.1

# Configuration and Formatting
python-dotenv>=0.19.0
typing-extensions>=4.0.0
toml>=0.10.2
ujson>=5.4.0
configparser>=5.3.0

# Development Tools
black>=22.3.0
isort>=5.10.1
mypy>=0.991
pylint>=2.15.0
pytest>=7.3.0
pytest-asyncio>=0.20.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
mkdocs>=1.4.0
mkdocs-material>=8.5.0

# Optional Dependencies
# Uncomment if needed
# ipython>=8.0.0  # For interactive shell features
# pygments>=2.13.0  # For syntax highlighting
# watchdog>=2.1.9  # For file system monitoring
# dbus-python>=1.2.18  # For system notifications